Seperate versions from models in data_classes.
