+++ tests/tests_integration/test_api/test_model_hosting/test_versions.py
@@ -6,7 +6,7 @@ import pytest
 
 from cognite.client._api.model_hosting.models import PredictionError
 from cognite.client._api.model_hosting.versions import EmptyArtifactsDirectory
-from cognite.client.data_classes.model_hosting.models import (
+from cognite.client.data_classes.model_hosting.versions import (
     ModelArtifact,
     ModelArtifactList,
     ModelVersion,
